add se agregan nuevas funciones
se agrega el apartado de RTN

+++ restaurante-laravel/database/migrations/2026_03_27_125054_create_reservaciones_table.php
@@ -8,18 +8,20 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::create('reservaciones', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('mesa_id')->constrained('mesas')->onDelete('cascade');
-            $table->string('cliente_nombre')->nullable();
-            $table->string('cliente_telefono')->nullable();
-            $table->date('fecha');
-            $table->time('hora');
-            $table->text('notas')->nullable();
-            $table->enum('estado', ['pendiente', 'completada', 'cancelada'])->default('pendiente');
-            $table->foreignId('usuario_id')->constrained('users');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('reservaciones')) {
+            Schema::create('reservaciones', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('mesa_id')->constrained('mesas')->onDelete('cascade');
+                $table->string('cliente_nombre')->nullable();
+                $table->string('cliente_telefono')->nullable();
+                $table->date('fecha');
+                $table->time('hora');
+                $table->text('notas')->nullable();
+                $table->enum('estado', ['pendiente', 'completada', 'cancelada'])->default('pendiente');
+                $table->foreignId('usuario_id')->constrained('users');
+                $table->timestamps();
+            });
+        }
     }
 
     public function down(): void
